samples/python/requirements.txt
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk
